+++ at_contacts_flutter/lib/at_contacts_flutter.dart
@@ -3,3 +3,4 @@ library at_contacts_flutter;
 export 'screens/contacts_screen.dart';
 export 'screens/blocked_screen.dart';
 export 'utils/init_contacts_service.dart';
+export 'desktop_screens/desktop_contacts_screen.dart';