+++ at_contacts_group_flutter/lib/models/group_contacts_model.dart
@@ -70,3 +70,4 @@ class GroupContactsModel {
 
 /// Enum for contact types
 enum ContactsType { CONTACT, GROUP }
+enum ContactTabs { RECENT, FAVS, ALL }