+++ at_onboarding_flutter/lib/services/size_config.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+/// This class helps with responsiveness in various screen sizes
 class SizeConfig {
   SizeConfig._();
 
@@ -55,6 +56,10 @@ class SizeConfig {
   }
 
   double getWidthRatio(double val) {
+    if (screenWidth! >= 1200) {
+      return val;
+    }
+
     double res = (val / refWidth) * 100;
     double temp = res * blockSizeHorizontal;
 
@@ -62,12 +67,20 @@ class SizeConfig {
   }
 
   double getHeightRatio(double val) {
+    if (screenWidth! >= 1200) {
+      return val;
+    }
+
     double res = (val / refHeight) * 100;
     double temp = res * blockSizeVertical;
     return temp;
   }
 
   double getFontRatio(double val) {
+    if (screenWidth! >= 1200) {
+      return val;
+    }
+
     double res = (val / refWidth) * 100;
     double temp = 0.0;
     if (screenWidth! < screenHeight) {