+++ at_backupkey_flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.1.1
+- **FIX**: Added file location picker for desktop platforms to save backup keys
+
+## 3.1.0
+- **FEAT**: Added example
+- **FEAT**: Added platform support for macos, windows and linux
+
 ## 3.0.1
 - added a public method to show a backup key dialog