+++ at_contacts_group_flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.1.1
+- **REFACTOR**: Use of SizeConfig from at_common_flutter
+- Updated package dependencies
+
+## 3.1.0
+- **FEAT**: Added platform support for macos, windows and linux
+
 ## 3.0.3
 - Updated dependencies and bug fixes